OneDriveExplorer/requirements.txt
ttkthemes
Pillow
pandas
numpy
pytz
python-registry
keyboard
openpyxl
pandastable
dissect.cstruct
psutil
pytsk3
quickxorhash